sistema organizado em diferentes arquivos e diretórios
