+++ src/main/java/pedido/Detalhe.java
@@ -1,0 +1,9 @@
+package pedido;
+
+public abstract class Detalhe implements Pedido {
+	protected Pedido pedido;
+
+	public Detalhe(Pedido pedido) {
+		this.pedido = pedido;
+	}
+}
