+++ src/main/java/pedido/Pedido.java
@@ -1,0 +1,6 @@
+package pedido;
+
+public interface Pedido {
+	public String getDescricao();
+	public double getCusto();
+}
